Initial GC with Cons, Vectors and Early Closures
A lot of incremental work has add up to significant changes.
There is initial simple -and in a couple ways unfinished- implementation
of GC. Cons and Vectors were implemented with some initial functions to
create and use them. Lambda/Closures are currently work in progress
(i.e. They will be created, but cannot really be called).
There are many small changes to environment, parser, and so on.
Generally to accomodate new compiler features, though a few fixes were
done.