New Variable System and Constants
Simple atoms that are not modified in the program are now turned into
constants instead of occupying precious register and stack space.
Naturally when variables are mutable they will be stored in registers
and stack including previous constants in cases like the interpreter.
This is enabled by a different system for the environment.

Boxing is broken as the free system has to be reworked into variables.
Currently WIP on quotes and better parser.