Free Variables now Handled in the Variable System
Recent update to the variable system was left unfinished in regard to
the lambda's free variables. This addresses that and integrates the free
variables into the env instead of being separate, which besides being
just a simpler system to follow, will also give the compiler better type
hints and other useful information.