Lambda Fixes and Parser Rewritten
A few more issues of free variables were considered and
I worked on more improvements of the system.

The largest change is the parser however.
The old version had a lot of issues and was not pleasant to work with.
The new version exceeded in feature parity and fixed a few common issues.
It is also a much clearer and nicer system to work on.